modularized map-setup.js
